Fixed some warnings and errors for Visual C++. For example, including winsock2.h after either utility.h or timer.h would result in a number of redefinition errors.

+++ array.h
@@ -300,7 +300,7 @@ struct array {
 	 * decrements array::length by `1`. The element initially at `index` is not
 	 * freed.
 	 */
-	void remove(unsigned int index)
+	void remove(size_t index)
 	{
 		core::move(data[length - 1], data[index]);
 		length--;
